Update sr2vi.R
The function processLandsat() gives an error when fmask masks all pixels in a given extent. An additional condition: "if(freq(mask,value=0) + freq(mask,value=1) == 0)) {vi[] <- NA}" should fix the problem (Fmask contains 0 and 1 values for land and water surface). In this way processLandsat() shouldn't stop which is particularly useful when running in batch mode for a large number of datasets.